autoflow-agent/requirements.txt
google-adk[database]==1.4.1
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
